Add reverse option for matcher
For the case when the most significant part is at the end of the
strings to be matched.

This is still not 100%, despite cranking the gray cells to the max. Turns out
that doing smart reverse boundary matching is quite a challenge with lpeg,
at least for yours truly. Or it's just a challenge regardless of the approach
used. Whatever the case, lpeg once again makes me feel stupid.