Revert "Run handlers in their own coroutines"
Alas, this is not still stable it seems. While having the Scintilla
integration in all FFI made it seem like it worked, I've now seen
some segfaults. It could be that this does not play well with lgi,
which is plain C api. Will have to save that for some rainy day and
see where to go with it, whether to skip lgi altogether and go full
FFI even for Gtk or investigate further. We don't need it right now
so it will have to wait.

This reverts commit c6b4cbcc96210d839cb9e5be0e364e98b71d3c04.