Start work on new ljglibs module
A potential standalone FFI binding to the g* library stack. Very
limited as it's only what we need.

First out, GFile and GRegex. Regular expressions were always FFI, so that
was just separating out glib from howl and cleaning things up. GFile was
new though. Although this was not done for performance reasons, it was nice
to see a noticeable speedup in startup and file-heavy parts such as project-open
for a larger project.