Trigger line number column width calculation later
As it was, it would end up calling into Scintilla when it's in a weird
state, leading to segfaults. Did not see it myself, but turned out that
the vi bundle masked the problem for me since a cursor style setting
call seemed to cause Scintilla to be properly initialized first.

Closes #3, as reported by @pinealservo.