Add howl.clipboard
Howl's own clipboard. Synchronizes with the system clipboard
as necessary, but lets us keep track of an arbitrary amount
of clips within Howl. Also allows for each clip to contain
additional metadata.

One use of this metadata is part of this commit - the ability
to specify that a certain clips contains whole lines, and
should be inserted as a line block upon paste.