Ensure that the readline only grows taller during a read()
Having the readline grow and shrink is annoying since it
requires the eyes to move up and down. Now the readline
will only grow during one read() invocation, and will
keep the same fixed size regardless of the amount of items
in the completion list.

Thanks to @shalabhc for the prodding.
 # Please enter the commit message for your changes. Lines starting