Remove fuzzy matching and fix boundary matching
The fuzzy matching was not useful, except for differentiating between
similar existing matches obtained by other means. At the same time
they caused a lot of uneccessary noice.

Removing the fuzzy matching led to rewriting matching in general, and
boundary matching in particular. While  it was an exciting idea to employ
lpeg for matching, it really didn't work out well for it. The concept
of building up lpeg patterns iteratively for a boundary match was mind
boggling in complexity. Also, the lack of backtracking meant it never
really worked that well (i.e. was broken for multiple cases). Regexes
proved a much better solution for this.