Rework the command module
It has long been the heart of darkness when it comes to
complexity in Howl. A big reason for this is that it's tried
to do a lot, particularly handling multiple inputs. Multiple
inputs has not actually been used once though, and the resulting
mess of trying to parse parameters and keep it accurate has only
led to a tangled mess.

Now it's simpler. A command can have one input, and that input will
receive all text as is inputted in the Readline. We're still supporting
the old table style .inputs for now though to provide som backwards
compatibility.