Allow looking up coordinates below visible area
As dubious as this might sound, it's likely a good thing. The
scenario is that the window might be temporarily resized when
the command line is shown, but as a result of a command something
needs to happen at a position that will be visible as soon as the
command line is hidden.