Auto adjust positions for offset calculations
Makes it similar to the current master. Held off on this for a while
to catch bad code, but it's probably time to let it go now. One of the
bigger cases for doing this is ending lines without trailing newlines,
where all callers would have to do explicit checks to avoid overstepping
by one. For the user API it's probably nice to be lenient as well.