Fix "lockup" problems with extremely large lines
As seen when opening a file with a minimized JS library on one line.
The problem is that we need to determine the height of the line (to
support variable height lines), and Pango apparently breaks down when
running the display calculations for the layout. We "fix" this by
not even trying to display super complex lines with correct styling,
but override the styling with the 'blob' style. Since this is not
strictly related to the lenght of the line but rather the complexity
of the styling that's what we check.