Add Buffer:change, supporting efficient composite changes
The big problem with doing a massive amount of updates for a buffer is
the corresponding amounts of notifications to views. This get's
very inefficient very fast. The new `change` method allows for a way
around this, by allowing a range to be modified with all changes
collected in one revision. In contrast to `as_one_undo`, only one
notification will be fired to listeners.