Switch ListWidget search higlighting to highlights
Instead of styling as before. The reason styling was used to begin
with was that the Scintilla indicators didn't support changing the
text color, and this is no longer the case with aullar. However, using
styling meant that bold styles could be used, and that's now impossible
since highlights can't change font properties except the colour, so
it means the bold is out.