Major theming and UI rework
This moves away from the old CSS based theming. CSS theming is still
used, but only for the status and indicator bars. All other elements
have been updated to to use the new ContentBox class extensively.

The major driver for for this is to allow for more advanced themes, and
allow things such as border radiuses, transparency everywhere, etc. The
currenty themes have been adapted for the new structure but have been
kept pretty much the same.