Improve determination of last line to style
There was previously a problem when editing an empty or short file and inserting
lots of line, resulting in not everything visible being styled. We're now
smarter about that. We're not super smart though, but it seems likely we're
smart enough that we won't have this problem any more in practice.