Make scrollbars themable
It didn't show itself to be much of a problem, but probably only since the
overlay scrollbars were often a visible orange, which wasn't that pleasing but
at least clearly visible across all themes. When trying out Elementary OS
however, with black scrollbars, it's apparent that there is a problem if the
scrollbars aren't visible.