Make sure we show any files opened with the --reuse parameter
Previously those were just added to the list, which meant that they
ended up at the bottom of the buffer list. Now we make sure that we display
any new files, filling out any existing editors with the new files.

Previous loading didn't do any existing check either, we we really need to
when files are opened with the --reuse parameter.

Fixes #213.