Bundle needed dependencies for running specs
This removes the dependency on luarocks for running Howl specs. This is
good in itself, but it's particularly good since we rely on the stable
busted and that's pretty much possible to install using luarocks now a
days.

The one cumberson system dependency from the busted chain, lfs, we
solve by implementing a crude lfs shim.