C/C++: Handle highlighting of partial type defs
When we encounter a partial type def we need to provide the knowledge
to the styler that it's incomplete, so that it can trigger a larger
refresh as needed. One example of this is as provided in #185, where
the type defining keyword is on one line with the rest is on the
subsequent line.

Fixes #185.