Move away from pango style CSS font declaration
Newer versions of Gtk apparently barfs on this.

Also, have to investigate what is going on with the font size specs in
general as we have to pass the font size as pixels for it to match up
with the rest. And this while the pango docs still talks about points.