Major config implementation refactor.
While buffer.config, mode.config and the set* commands can still be
used the same way, the underlying implementation has been changed.

The config is now organized along two dimensions - a primary dimension
called scope which is a hierarchial namespace, and a secondary
dimension called layer which can be any string. When defined, a layer
may be associated with another layer as the parent. For each scope,
multiple layers may exist and for each layer within the scope multiple
name/value pairs may exist.

Values are looked up in a scope, and if not found, the parent scopes
are checked, all the way from the immedidate parent to the global
scope, represented by the emtpy string ''. For each scope, value
resolution may check multiple layers.

Proxy config objects on buffer and mode hide most of this machinery by
exposing config values as attributes.