Automatic configuration persistence
Config set via command line or API is now automatically persisted on
clean exit. Exiting via Ctrl-C does not save the config. The saved
config is loaded automatically on next startup.

The 'persist_config' configuration value determines if the config will
be persisted or not. It can be set globally, or per file using
'set-for-buffer'.

Config set via .howl/init.moon will overwrite the loaded config, and
will be saved on next exit.