A new 'set' command that replaces all existing 'set*' commands
The new set command has the following form:

  set <name>@<scope>[<layer>]=<value>

The [<layer>] part is optional.

Examples:

  set theme@global=Monokai
  set indent@global[mode:moonscript]=2
  set auto_inspect@buffer=manual

Selection lists provide auto-complete for the <name> and <scope> part,
making it generally unnecessary to type out the whole thing.

This is now implemented as a sequence of three interactions, rather
than one complex text entry interaction.

Backspacing backtracks to the previous interaction, allowing you to
edit any part of the command, but different than the current behavior
in that it clears the whole previous selection, rather than deleting
one character. (In that sense it is similar to backspacing into the
parent directory in the file browser).