Allowing running different inspectors for idle and save
The previous solution had only one list of inspectors, and you could
chose when those would run (idle, save or manually). While this was a
simpler (and thus nicer in that sense) solution, it didn't allow for
running lightweight inspections on idle and heavy weight inspections
on save.

With this split there are now two configuration variables for
controlling what inspectors are run, `inspectors_on_idle` and
`inspectors_on_save`. The `auto_inspect` option now allows the
following three alternatives for controlling the running behaviour:

- manual: Run all inspectors when explicitly asked
- idle: Run on-idle inspectors on idle and on-save inspectors on save
- save: Run all inspectors, but only on save