Removed whitespace and reverted c<motion> change for end of line
Misinterpreted the spec for c<motion> at the end of a line to be confined to that line.
Really the spec intends the last non-blank character, for which the existing code was fine.
Nothing is mentioned about handling actual end of line behavior any differently, so I'm leaving it alone.
Undo!