Mano a mano with Gtk scrollbars again
So far it's been a losing battle, but perhaps this time? The last change
to get rid of the pesky Gtk warning resulted in erratic scroll behaviour
instead due to this Gtk issue:

https://bugzilla.gnome.org/show_bug.cgi?id=765410

Apparently we cannot get syncing of scrollbars to work at all from within
the size allocate handler. Since none of the other queueing mechanisms
seems to have an effect (queue_resize, etc.) we just bend over and add
an ugly work-around based on triggering an additional refresh after the
size allocate is done.

For issue #371.