Improve Python lexing of f strings
The previous lexing was too loose. It would start a f string, scan until the
end or until a starting brace, and then lex anything followed by more f string
scanning. This obviously causes problems if the string has ended, but we still
lexed something else and continue with the f string.

Instead tighten up the lexing. We scan the f string until the end or until the
first brace. If and only if we're looking at an interpolation then attempt that,
but otherwise handle ends, etc., correctly.

Also lex double braces in f strings correctly.

Closes #379.