Add howl.activities
The activities module keeps track of long running operations with Howl, that
either is or should appear to be blocking in nature.

While most operations within Howl are fast enough that any blocking is at most a
nuisance, some things can be slow enough that it’s not reasonable to perform
them in a blocking manner without any feedback to the user (apart from an
unresponsive editor). Alternatively, some operations can be slow but
non-blocking (e.g. using some of Howl’s asynchronous APIs), and could by mistake
allow the user to work before the result of a needed operation is ready.

Using the run or run_process functions available from the activities module,
code can run designated blocks of code as activities. An activity is in its most
basic form simply a named part of code, with an associated dynamic status, whose
execution is supervised by the activities module. If the execution of the
activity has not completed within a specific interval the user is alerted to the
ongoing operation via a popup at the bottom at the current window, providing
insight into what is being done and optionally any progress.