select_location interaction no longer leaves any highlights
It's not theoretically sound to begin with since there's no telling
what will happen with the returned location, which makes any lingering
highlights a bad side effect.

On top of that it, for natural reasons, only left highlights on currently
open buffers, whereas with non-opened buffers it could not. That made it
inconsistent.

Now every caller has to explicitly highlight any are of interest themselves.