Add Application.open(..)
In contrast to `open_file`, which only opens file `open` can open the
more general 'locations'. This can a position in a file as well, but can
also point to a buffer just as well. The exact position of the cursor can
be specified in multiple ways, and highlights can optionally be given as well.