@feature: "cargo check"-inspector for the rust bundle
This adds an "on_idle" inspector in the rust bundle, that calls
`cargo check` and parses the output json for warning and errors.
It was necessary to extend `inspect.moon` to accept options for
enabling/disabling stdout/stderr because cargo sends the json
to stdout and some other non-json to stderr.
This produced a bug in `activities.moon` where the pump-methods
of the process always received a `on_stdout` and a `on_stderr`
handler, regardless of the existence of a `stdout` or `stderr`
object respectivly in the process.
The fix simply checks for the existence of these objects and
omits passing the respective handler if the output stream does
not exist.