Add breadcrumb_tolerance config setting
This causes breadcrumbs within the specified limit to be merged when
added, and when navigating to be skipped over if the cursor is close
enough.

This allows us to more liberally drop breadcrumbs where we want them,
for instance after doing searches we now drop a new crumb at the
moved-to location. If we had done so previously, an instant back
navigation would have meant not moving the cursor at all.

The limit is set to 10 by default now, but can be changed if experience
says something else would be better.