Start reworking memory management vs Gtk
We don't have destroy any more, and the release thing isn't really good so
time to spend some effort making everything collectable. Most notably, the problem
is with Gtk/C callbacks, that needs a reference to the matching lua instance.
We now tie callbacks to the life cycle of the lua ref, and make it the
responsibility of the called to ensure liveness.