Fighting with flairs again (last time?)
Problem is with flairs covering a range of variable size character, where we need
to position the drawing correctly. One would think that given the starting
character and size differences for the range one could deduce the correct Y
positioning for whole range, but so far no. For now we get the biggest character in
the range and use that. Not optimal, but am attempting to escape with some of my
sanity intact.