Pixel hunting
First off set the assumption that if line 1 has height 10, then line 2 will start
at 11. Also clip every display line before drawing. There's an issue with cairo
showing itself for embedded backgrounds so this helps that. It's really a more
complex issue with the pixel alignment with cairo, etc., but better for now.