Give up on cairo and extents
For *some* reason, even when an area is cleared and later filled with the same
rectangle we still end up with with display artifacts.. Don't have time for this
shit. We now draw all flairs with the naive rects, i.e. plain y and height. For the
one case where we care (background ranges) we use a full_height flag to increase the
rect to fill everything.